refactor grpc: cleanup client CallState commit_hash:005f29f09e48e35ade73a85a33f68094fc8ca3ec
